feat: Mayor Monster Hunt quest, sword/equipment/attack, tabbed quick screen
Milestone A (health pass):
- make lint: compile-to-assembly -Wf-Wall pass (the normal build can't use
  -Wall: sdcc --use-stdout leaks warnings into the .asm stream).  Lint caught
  and fixed a real bug: progression thresholds were uint8_t truncating
  values > 255 (now uint16_t).  lint reports zero warnings.
- Removed dead party_add_member.

Milestone B (quest + equipment + quick screen):
- Event engine generalized: EventCond list (flag is/not set, variable >=/==),
  EVENT_ACTION_ADD_ITEM, EVENT_TRIGGER_ACTOR_DEFEATED; world_on_battle_end
  now takes Game* and routes defeated-hostile progress through the event
  table.  Quest state/objective are generic variables
  (QUEST_MONSTER_HUNT 0/1/2, MONSTERS_REMAINING) owned by events.
- Events: QUEST_START (mayor intro starts quest, sets MET_MAYOR + 3 monsters),
  QUEST_ACTIVE, QUEST_COMPLETE (gives SWORD exactly once), QUEST_DONE,
  MONSTER_DEFEATED (decrement while ACTIVE).
- Sword/equipment/stats: ITEM_SWORD (+3), ItemKind, EquipmentState in
  GameState, game_hero_attack = 3 + weapon bonus, Combatant.attack,
  battle_start carries player attack; item_equip + ITEM_EQUIPPED telemetry.
- Quick screen (SCREEN_ITEM) is now a tabbed menu ITEM|EQUIP|STATUS with an
  HP/GOLD header; START opens it (overworld SELECT does nothing), SELECT
  focuses the tab row, arrows + A confirm, B closes; battle SELECT opens the
  ITEM tab.  Standalone STATUS screen removed.
- Training slime: equipping the sword spawns a fresh non-persistent hostile
  slime (world_spawn_actor + game_on_equip) to test the one-shot kill.
- Wire contract v0x03: equipment byte added to descriptor + snapshot
  (ROM+host in sync, roundtrip passes).
- Harness: equip_item action, context-aware use_item (START overworld /
  SELECT battle), attack assertion, EQUIPMENT/ATTACK in the semantic dump.
- Scenarios: 9 new (quest_start/progress/complete_interact/no_double_reward/
  gate, equip_sword, attack_base, battle_sword_damage, quick_equip_ui);
  migrated mayor_reinteract (QUEST_ACTIVE) and status_screen (STATUS tab).

Harness: 73/73 scenarios pass. make lint: no warnings. make test release
validation passes.

+++ src/rpg/progression.c
@@ -5,9 +5,9 @@
 /* ── Static progression definitions (rules), separate from state ─────
  * Fixed lookup tables (Game Boy friendly).  Threshold entries are the
  * progress needed to advance FROM level N to level N+1. */
-static const uint8_t g_hero_thresholds[] = {100, 250, 450, 700, 1000, 1400, 1900, 2500, 3200, 4000};
-static const uint8_t g_weapon_thresholds[] = {50, 120, 220, 350, 520, 750, 1020};
-static const uint8_t g_companion_thresholds[] = {60, 150, 280, 450, 660, 900, 1200};
+static const uint16_t g_hero_thresholds[] = {100, 250, 450, 700, 1000, 1400, 1900, 2500, 3200, 4000};
+static const uint16_t g_weapon_thresholds[] = {50, 120, 220, 350, 520, 750, 1020};
+static const uint16_t g_companion_thresholds[] = {60, 150, 280, 450, 660, 900, 1200};
 
 typedef struct {
     uint8_t type;
